1. Allowing boolean objects to be set as NULL. This is necessary eg for unsetting boolean form fields when the database supports NULL values.
important notice: boolean properties of new objects will no longer be
able to be left unset if the correlating database field is configured as
"NOT NULL". Setting the 'default' property for all such fields will fix
this - or if you ensure that the property is set before the first time
the object is saved (eg via a form field).

2. removing ineffectual methods for "RequiredAdminFormFields"

3. minor styles including spaces->tabs for internal consistency